restructure folders and rename files
