+++ .gitignore
@@ -1,5 +1,6 @@
 node_modules/
 dist
+build
 !./dist/.gitkeep
 *.zip
 .DS_Store
