+++ shells/dev/alias.js
@@ -2,6 +2,6 @@
 const path = require("path");
 
 module.exports = {
-  src: path.resolve(__dirname, "../src"),
+  src: path.resolve(__dirname, "../../src"),
   "react-native": "react-native-web",
 };
