+++ shells/dev/webpack.config.js
@@ -1,6 +1,6 @@
 const path = require("path");
 const webpack = require("webpack");
-const alias = require("../alias");
+const alias = require("./alias");
 const UglifyPlugin = require("uglifyjs-webpack-plugin");
 const BundleAnalyzerPlugin = require("webpack-bundle-analyzer")
   .BundleAnalyzerPlugin;
