--- shells/webextension/stats.json
@@ -1,201 +1,0 @@
-{
-  "errors": [],
-  "warnings": [],
-  "version": "1.15.0",
-  "hash": "1f76099fd3ef7e0bf220",
-  "time": 17331,
-  "publicPath": "",
-  "assetsByChunkName": {
-    "backend": "backend.js",
-    "background": "background.js",
-    "devtools": "devtools.js",
-    "devtools-background": "devtools-background.js",
-    "hook": "hook.js",
-    "proxy": "proxy.js"
-  },
-  "assets": [
-    {
-      "name": "backend.js",
-      "size": 57385,
-      "chunks": [0],
-      "chunkNames": ["backend"],
-      "emitted": true
-    },
-    {
-      "name": "background.js",
-      "size": 1224,
-      "chunks": [1],
-      "chunkNames": ["background"],
-      "emitted": true
-    },
-    {
-      "name": "devtools.js",
-      "size": 962432,
-      "chunks": [2],
-      "chunkNames": ["devtools"],
-      "emitted": true
-    },
-    {
-      "name": "devtools-background.js",
-      "size": 796,
-      "chunks": [3],
-      "chunkNames": ["devtools-background"],
-      "emitted": true
-    },
-    {
-      "name": "hook.js",
-      "size": 1520,
-      "chunks": [4],
-      "chunkNames": ["hook"],
-      "emitted": true
-    },
-    {
-      "name": "proxy.js",
-      "size": 621,
-      "chunks": [5],
-      "chunkNames": ["proxy"],
-      "emitted": true
-    }
-  ],
-  "chunks": [
-    {
-      "id": 0,
-      "rendered": true,
-      "initial": true,
-      "entry": true,
-      "extraAsync": false,
-      "size": 149118,
-      "names": ["backend"],
-      "files": ["backend.js"],
-      "hash": "cebe6658b9d4fed0088c",
-      "parents": [],
-      "origins": [
-        {
-          "moduleId": 0,
-          "module": "/mnt/c/Users/james/Projects/apollo/apollo-client-devtools/node_modules/babel-loader/lib/index.js!/mnt/c/Users/james/Projects/apollo/apollo-client-devtools/shells/webextension/src/backend.js",
-          "moduleIdentifier": "/mnt/c/Users/james/Projects/apollo/apollo-client-devtools/node_modules/babel-loader/lib/index.js!/mnt/c/Users/james/Projects/apollo/apollo-client-devtools/shells/webextension/src/backend.js",
-          "moduleName": "./src/backend.js",
-          "loc": "",
-          "name": "backend",
-          "reasons": []
-        }
-      ]
-    },
-    {
-      "id": 1,
-      "rendered": true,
-      "initial": true,
-      "entry": true,
-      "extraAsync": false,
-      "size": 2236,
-      "names": ["background"],
-      "files": ["background.js"],
-      "hash": "9a906d36870b6110cdad",
-      "parents": [],
-      "origins": [
-        {
-          "moduleId": 0,
-          "module": "/mnt/c/Users/james/Projects/apollo/apollo-client-devtools/node_modules/babel-loader/lib/index.js!/mnt/c/Users/james/Projects/apollo/apollo-client-devtools/shells/webextension/src/background.js",
-          "moduleIdentifier": "/mnt/c/Users/james/Projects/apollo/apollo-client-devtools/node_modules/babel-loader/lib/index.js!/mnt/c/Users/james/Projects/apollo/apollo-client-devtools/shells/webextension/src/background.js",
-          "moduleName": "./src/background.js",
-          "loc": "",
-          "name": "background",
-          "reasons": []
-        }
-      ]
-    },
-    {
-      "id": 2,
-      "rendered": true,
-      "initial": true,
-      "entry": true,
-      "extraAsync": false,
-      "size": 2284059,
-      "names": ["devtools"],
-      "files": ["devtools.js"],
-      "hash": "de1433ab24f8eab92104",
-      "parents": [],
-      "origins": [
-        {
-          "moduleId": 0,
-          "module": "/mnt/c/Users/james/Projects/apollo/apollo-client-devtools/node_modules/babel-loader/lib/index.js!/mnt/c/Users/james/Projects/apollo/apollo-client-devtools/shells/webextension/src/devtools.js",
-          "moduleIdentifier": "/mnt/c/Users/james/Projects/apollo/apollo-client-devtools/node_modules/babel-loader/lib/index.js!/mnt/c/Users/james/Projects/apollo/apollo-client-devtools/shells/webextension/src/devtools.js",
-          "moduleName": "./src/devtools.js",
-          "loc": "",
-          "name": "devtools",
-          "reasons": []
-        }
-      ]
-    },
-    {
-      "id": 3,
-      "rendered": true,
-      "initial": true,
-      "entry": true,
-      "extraAsync": false,
-      "size": 2092,
-      "names": ["devtools-background"],
-      "files": ["devtools-background.js"],
-      "hash": "b071534ff47b58c0351d",
-      "parents": [],
-      "origins": [
-        {
-          "moduleId": 0,
-          "module": "/mnt/c/Users/james/Projects/apollo/apollo-client-devtools/node_modules/babel-loader/lib/index.js!/mnt/c/Users/james/Projects/apollo/apollo-client-devtools/shells/webextension/src/devtools-background.js",
-          "moduleIdentifier": "/mnt/c/Users/james/Projects/apollo/apollo-client-devtools/node_modules/babel-loader/lib/index.js!/mnt/c/Users/james/Projects/apollo/apollo-client-devtools/shells/webextension/src/devtools-background.js",
-          "moduleName": "./src/devtools-background.js",
-          "loc": "",
-          "name": "devtools-background",
-          "reasons": []
-        }
-      ]
-    },
-    {
-      "id": 4,
-      "rendered": true,
-      "initial": true,
-      "entry": true,
-      "extraAsync": false,
-      "size": 3520,
-      "names": ["hook"],
-      "files": ["hook.js"],
-      "hash": "baab6f55d7308b17db8c",
-      "parents": [],
-      "origins": [
-        {
-          "moduleId": 0,
-          "module": "/mnt/c/Users/james/Projects/apollo/apollo-client-devtools/node_modules/babel-loader/lib/index.js!/mnt/c/Users/james/Projects/apollo/apollo-client-devtools/shells/webextension/src/hook.js",
-          "moduleIdentifier": "/mnt/c/Users/james/Projects/apollo/apollo-client-devtools/node_modules/babel-loader/lib/index.js!/mnt/c/Users/james/Projects/apollo/apollo-client-devtools/shells/webextension/src/hook.js",
-          "moduleName": "./src/hook.js",
-          "loc": "",
-          "name": "hook",
-          "reasons": []
-        }
-      ]
-    },
-    {
-      "id": 5,
-      "rendered": true,
-      "initial": true,
-      "entry": true,
-      "extraAsync": false,
-      "size": 932,
-      "names": ["proxy"],
-      "files": ["proxy.js"],
-      "hash": "8edefce77cbc4f78b44f",
-      "parents": [],
-      "origins": [
-        {
-          "moduleId": 0,
-          "module": "/mnt/c/Users/james/Projects/apollo/apollo-client-devtools/node_modules/babel-loader/lib/index.js!/mnt/c/Users/james/Projects/apollo/apollo-client-devtools/shells/webextension/src/proxy.js",
-          "moduleIdentifier": "/mnt/c/Users/james/Projects/apollo/apollo-client-devtools/node_modules/babel-loader/lib/index.js!/mnt/c/Users/james/Projects/apollo/apollo-client-devtools/shells/webextension/src/proxy.js",
-          "moduleName": "./src/proxy.js",
-          "loc": "",
-          "name": "proxy",
-          "reasons": []
-        }
-      ]
-    }
-  ],
-  "children": []
-}
