+++ src/ChromeStorageAdapter.js
