+++ src/application/components/Explorer/graphiql-overrides.less
