+++ src/application/components/Explorer/graphiql.less
