+++ src/application/components/Images/Warning.js
