+++ src/application/components/Inspector/Inspector.js
