+++ src/application/components/Inspector/index.js
