+++ src/application/components/Inspector/inspector.less
