+++ src/application/components/Logger/Logger.js
