+++ src/application/components/Logger/index.js
