+++ src/application/components/Panel.js
