+++ src/application/components/WatchedQueries/WatchedQueries.less
