+++ src/application/components/bridge/index.js
