+++ src/application/style.less
@@ -10,7 +10,7 @@ body,
   padding: 0;
   margin: 0;
   font-size: 12px;
-  color: #333333;
+  color: red;
   overflow: hidden;
 }
