+++ src/extension/background/background.js
