+++ src/extension/devtools/devtools.html
