+++ src/extension/devtools/panel.html
