+++ src/extension/devtools/panel.js
@@ -1,8 +1,8 @@
 // this script is called when the Apollo devtools panel
 // is activated
-import wat, { initDevTools } from "src/devtools";
-import Bridge from "src/bridge";
-import { createChromeStorageAdapter } from "./ChromeStorageAdapter";
+import { initDevTools } from "../../application";
+import Bridge from "../../bridge";
+import { createChromeStorageAdapter } from "../../ChromeStorageAdapter";
 
 createChromeStorageAdapter(chrome.storage.local, storage => {
   initDevTools({
