+++ src/extension/manifest.json
