+++ src/extension/tab/hook.js
